+++ crates/polars-plan/src/logical_plan/conversion/dsl_to_ir.rs
@@ -1,11 +1,4 @@
-use std::borrow::Cow;
-
-use polars_core::prelude::*;
-use polars_utils::vec::ConvertVec;
-use recursive::recursive;
-
-use crate::constants::get_len_name;
-use crate::prelude::*;
+use super::*;
 
 pub fn to_expr_ir(expr: Expr, arena: &mut Arena<AExpr>) -> ExprIR {
     let mut state = ConversionState::new();
@@ -66,8 +59,8 @@ fn to_aexprs(input: Vec<Expr>, arena: &mut Arena<AExpr>, state: &mut ConversionS
 }
 
 fn set_function_output_name<F>(e: &[ExprIR], state: &mut ConversionState, function_fmt: F)
-where
-    F: FnOnce() -> Cow<'static, str>,
+    where
+        F: FnOnce() -> Cow<'static, str>,
 {
     if state.output_name.is_none() {
         if e.is_empty() {
@@ -510,502 +503,4 @@ pub fn to_alp(
         },
     };
     Ok(lp_arena.add(v))
-}
-
-/// converts a node from the AExpr arena to Expr
-#[recursive]
-pub fn node_to_expr(node: Node, expr_arena: &Arena<AExpr>) -> Expr {
-    let expr = expr_arena.get(node).clone();
-
-    match expr {
-        AExpr::Explode(node) => Expr::Explode(Arc::new(node_to_expr(node, expr_arena))),
-        AExpr::Alias(expr, name) => {
-            let exp = node_to_expr(expr, expr_arena);
-            Expr::Alias(Arc::new(exp), name)
-        },
-        AExpr::Column(a) => Expr::Column(a),
-        AExpr::Literal(s) => Expr::Literal(s),
-        AExpr::BinaryExpr { left, op, right } => {
-            let l = node_to_expr(left, expr_arena);
-            let r = node_to_expr(right, expr_arena);
-            Expr::BinaryExpr {
-                left: Arc::new(l),
-                op,
-                right: Arc::new(r),
-            }
-        },
-        AExpr::Cast {
-            expr,
-            data_type,
-            strict,
-        } => {
-            let exp = node_to_expr(expr, expr_arena);
-            Expr::Cast {
-                expr: Arc::new(exp),
-                data_type,
-                strict,
-            }
-        },
-        AExpr::Sort { expr, options } => {
-            let exp = node_to_expr(expr, expr_arena);
-            Expr::Sort {
-                expr: Arc::new(exp),
-                options,
-            }
-        },
-        AExpr::Gather {
-            expr,
-            idx,
-            returns_scalar,
-        } => {
-            let expr = node_to_expr(expr, expr_arena);
-            let idx = node_to_expr(idx, expr_arena);
-            Expr::Gather {
-                expr: Arc::new(expr),
-                idx: Arc::new(idx),
-                returns_scalar,
-            }
-        },
-        AExpr::SortBy {
-            expr,
-            by,
-            sort_options,
-        } => {
-            let expr = node_to_expr(expr, expr_arena);
-            let by = by
-                .iter()
-                .map(|node| node_to_expr(*node, expr_arena))
-                .collect();
-            Expr::SortBy {
-                expr: Arc::new(expr),
-                by,
-                sort_options,
-            }
-        },
-        AExpr::Filter { input, by } => {
-            let input = node_to_expr(input, expr_arena);
-            let by = node_to_expr(by, expr_arena);
-            Expr::Filter {
-                input: Arc::new(input),
-                by: Arc::new(by),
-            }
-        },
-        AExpr::Agg(agg) => match agg {
-            AAggExpr::Min {
-                input,
-                propagate_nans,
-            } => {
-                let exp = node_to_expr(input, expr_arena);
-                AggExpr::Min {
-                    input: Arc::new(exp),
-                    propagate_nans,
-                }
-                .into()
-            },
-            AAggExpr::Max {
-                input,
-                propagate_nans,
-            } => {
-                let exp = node_to_expr(input, expr_arena);
-                AggExpr::Max {
-                    input: Arc::new(exp),
-                    propagate_nans,
-                }
-                .into()
-            },
-
-            AAggExpr::Median(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::Median(Arc::new(exp)).into()
-            },
-            AAggExpr::NUnique(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::NUnique(Arc::new(exp)).into()
-            },
-            AAggExpr::First(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::First(Arc::new(exp)).into()
-            },
-            AAggExpr::Last(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::Last(Arc::new(exp)).into()
-            },
-            AAggExpr::Mean(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::Mean(Arc::new(exp)).into()
-            },
-            AAggExpr::Implode(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::Implode(Arc::new(exp)).into()
-            },
-            AAggExpr::Quantile {
-                expr,
-                quantile,
-                interpol,
-            } => {
-                let expr = node_to_expr(expr, expr_arena);
-                let quantile = node_to_expr(quantile, expr_arena);
-                AggExpr::Quantile {
-                    expr: Arc::new(expr),
-                    quantile: Arc::new(quantile),
-                    interpol,
-                }
-                .into()
-            },
-            AAggExpr::Sum(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::Sum(Arc::new(exp)).into()
-            },
-            AAggExpr::Std(expr, ddof) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::Std(Arc::new(exp), ddof).into()
-            },
-            AAggExpr::Var(expr, ddof) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::Var(Arc::new(exp), ddof).into()
-            },
-            AAggExpr::AggGroups(expr) => {
-                let exp = node_to_expr(expr, expr_arena);
-                AggExpr::AggGroups(Arc::new(exp)).into()
-            },
-            AAggExpr::Count(expr, include_nulls) => {
-                let expr = node_to_expr(expr, expr_arena);
-                AggExpr::Count(Arc::new(expr), include_nulls).into()
-            },
-        },
-        AExpr::Ternary {
-            predicate,
-            truthy,
-            falsy,
-        } => {
-            let p = node_to_expr(predicate, expr_arena);
-            let t = node_to_expr(truthy, expr_arena);
-            let f = node_to_expr(falsy, expr_arena);
-
-            Expr::Ternary {
-                predicate: Arc::new(p),
-                truthy: Arc::new(t),
-                falsy: Arc::new(f),
-            }
-        },
-        AExpr::AnonymousFunction {
-            input,
-            function,
-            output_type,
-            options,
-        } => Expr::AnonymousFunction {
-            input: expr_irs_to_exprs(input, expr_arena),
-            function,
-            output_type,
-            options,
-        },
-        AExpr::Function {
-            input,
-            function,
-            options,
-        } => Expr::Function {
-            input: expr_irs_to_exprs(input, expr_arena),
-            function,
-            options,
-        },
-        AExpr::Window {
-            function,
-            partition_by,
-            options,
-        } => {
-            let function = Arc::new(node_to_expr(function, expr_arena));
-            let partition_by = nodes_to_exprs(&partition_by, expr_arena);
-            Expr::Window {
-                function,
-                partition_by,
-                options,
-            }
-        },
-        AExpr::Slice {
-            input,
-            offset,
-            length,
-        } => Expr::Slice {
-            input: Arc::new(node_to_expr(input, expr_arena)),
-            offset: Arc::new(node_to_expr(offset, expr_arena)),
-            length: Arc::new(node_to_expr(length, expr_arena)),
-        },
-        AExpr::Len => Expr::Len,
-        AExpr::Nth(i) => Expr::Nth(i),
-        AExpr::Wildcard => Expr::Wildcard,
-    }
-}
-
-fn nodes_to_exprs(nodes: &[Node], expr_arena: &Arena<AExpr>) -> Vec<Expr> {
-    nodes.iter().map(|n| node_to_expr(*n, expr_arena)).collect()
-}
-
-fn expr_irs_to_exprs(expr_irs: Vec<ExprIR>, expr_arena: &Arena<AExpr>) -> Vec<Expr> {
-    expr_irs.convert_owned(|e| e.to_expr(expr_arena))
-}
-
-impl IR {
-    #[recursive]
-    fn into_lp<F, LPA>(
-        self,
-        conversion_fn: &F,
-        lp_arena: &mut LPA,
-        expr_arena: &Arena<AExpr>,
-    ) -> LogicalPlan
-    where
-        F: Fn(Node, &mut LPA) -> IR,
-    {
-        let lp = self;
-        let convert_to_lp = |node: Node, lp_arena: &mut LPA| {
-            conversion_fn(node, lp_arena).into_lp(conversion_fn, lp_arena, expr_arena)
-        };
-        match lp {
-            IR::Scan {
-                paths,
-                file_info,
-                predicate,
-                scan_type,
-                output_schema: _,
-                file_options: options,
-            } => LogicalPlan::Scan {
-                paths,
-                file_info,
-                predicate: predicate.map(|e| e.to_expr(expr_arena)),
-                scan_type,
-                file_options: options,
-            },
-            #[cfg(feature = "python")]
-            IR::PythonScan { options, .. } => LogicalPlan::PythonScan { options },
-            IR::Union { inputs, options } => {
-                let inputs = inputs
-                    .into_iter()
-                    .map(|node| convert_to_lp(node, lp_arena))
-                    .collect();
-                LogicalPlan::Union { inputs, options }
-            },
-            IR::HConcat {
-                inputs,
-                schema,
-                options,
-            } => {
-                let inputs = inputs
-                    .into_iter()
-                    .map(|node| convert_to_lp(node, lp_arena))
-                    .collect();
-                LogicalPlan::HConcat {
-                    inputs,
-                    schema: schema.clone(),
-                    options,
-                }
-            },
-            IR::Slice { input, offset, len } => {
-                let lp = convert_to_lp(input, lp_arena);
-                LogicalPlan::Slice {
-                    input: Arc::new(lp),
-                    offset,
-                    len,
-                }
-            },
-            IR::Filter { input, predicate } => {
-                let lp = convert_to_lp(input, lp_arena);
-                let predicate = predicate.to_expr(expr_arena);
-                LogicalPlan::Filter {
-                    input: Arc::new(lp),
-                    predicate,
-                }
-            },
-            IR::DataFrameScan {
-                df,
-                schema,
-                output_schema,
-                projection,
-                selection,
-            } => LogicalPlan::DataFrameScan {
-                df,
-                schema,
-                output_schema,
-                projection,
-                selection: selection.map(|e| e.to_expr(expr_arena)),
-            },
-            IR::Select {
-                expr,
-                input,
-                schema,
-                options,
-            } => {
-                let i = convert_to_lp(input, lp_arena);
-                let expr = expr_irs_to_exprs(expr.all_exprs(), expr_arena);
-                LogicalPlan::Select {
-                    expr,
-                    input: Arc::new(i),
-                    schema,
-                    options,
-                }
-            },
-            IR::SimpleProjection { input, columns, .. } => {
-                let input = convert_to_lp(input, lp_arena);
-                let expr = columns
-                    .iter_names()
-                    .map(|name| Expr::Column(ColumnName::from(name.as_str())))
-                    .collect::<Vec<_>>();
-                LogicalPlan::Select {
-                    expr,
-                    input: Arc::new(input),
-                    schema: columns.clone(),
-                    options: Default::default(),
-                }
-            },
-            IR::Sort {
-                input,
-                by_column,
-                slice,
-                sort_options,
-            } => {
-                let input = Arc::new(convert_to_lp(input, lp_arena));
-                let by_column = expr_irs_to_exprs(by_column, expr_arena);
-                LogicalPlan::Sort {
-                    input,
-                    by_column,
-                    slice,
-                    sort_options,
-                }
-            },
-            IR::Cache {
-                input,
-                id,
-                cache_hits,
-            } => {
-                let input = Arc::new(convert_to_lp(input, lp_arena));
-                LogicalPlan::Cache {
-                    input,
-                    id,
-                    cache_hits,
-                }
-            },
-            IR::GroupBy {
-                input,
-                keys,
-                aggs,
-                schema,
-                apply,
-                maintain_order,
-                options: dynamic_options,
-            } => {
-                let i = convert_to_lp(input, lp_arena);
-                let keys = Arc::new(expr_irs_to_exprs(keys, expr_arena));
-                let aggs = expr_irs_to_exprs(aggs, expr_arena);
-
-                LogicalPlan::GroupBy {
-                    input: Arc::new(i),
-                    keys,
-                    aggs,
-                    schema,
-                    apply,
-                    maintain_order,
-                    options: dynamic_options,
-                }
-            },
-            IR::Join {
-                input_left,
-                input_right,
-                schema,
-                left_on,
-                right_on,
-                options,
-            } => {
-                let i_l = convert_to_lp(input_left, lp_arena);
-                let i_r = convert_to_lp(input_right, lp_arena);
-
-                let left_on = expr_irs_to_exprs(left_on, expr_arena);
-                let right_on = expr_irs_to_exprs(right_on, expr_arena);
-
-                LogicalPlan::Join {
-                    input_left: Arc::new(i_l),
-                    input_right: Arc::new(i_r),
-                    schema,
-                    left_on,
-                    right_on,
-                    options,
-                }
-            },
-            IR::HStack {
-                input,
-                exprs,
-                schema,
-                options,
-            } => {
-                let i = convert_to_lp(input, lp_arena);
-                let exprs = expr_irs_to_exprs(exprs.all_exprs(), expr_arena);
-
-                LogicalPlan::HStack {
-                    input: Arc::new(i),
-                    exprs,
-                    schema,
-                    options,
-                }
-            },
-            IR::Distinct { input, options } => {
-                let i = convert_to_lp(input, lp_arena);
-                LogicalPlan::Distinct {
-                    input: Arc::new(i),
-                    options,
-                }
-            },
-            IR::MapFunction { input, function } => {
-                let input = Arc::new(convert_to_lp(input, lp_arena));
-                LogicalPlan::MapFunction { input, function }
-            },
-            IR::ExtContext {
-                input,
-                contexts,
-                schema,
-            } => {
-                let input = Arc::new(convert_to_lp(input, lp_arena));
-                let contexts = contexts
-                    .into_iter()
-                    .map(|node| convert_to_lp(node, lp_arena))
-                    .collect();
-                LogicalPlan::ExtContext {
-                    input,
-                    contexts,
-                    schema,
-                }
-            },
-            IR::Sink { input, payload } => {
-                let input = Arc::new(convert_to_lp(input, lp_arena));
-                LogicalPlan::Sink { input, payload }
-            },
-            IR::Invalid => unreachable!(),
-        }
-    }
-}
-
-pub fn node_to_lp_cloned(
-    node: Node,
-    expr_arena: &Arena<AExpr>,
-    mut lp_arena: &Arena<IR>,
-) -> LogicalPlan {
-    // we borrow again mutably only to make the types happy
-    // we want to initialize `to_lp` from a mutable and a immutable lp_arena
-    // by borrowing an immutable mutably, we still are immutable down the line.
-    let alp = lp_arena.get(node).clone();
-    alp.into_lp(
-        &|node, lp_arena: &mut &Arena<IR>| lp_arena.get(node).clone(),
-        &mut lp_arena,
-        expr_arena,
-    )
-}
-
-/// converts a node from the IR arena to a LogicalPlan
-pub fn node_to_lp(node: Node, expr_arena: &Arena<AExpr>, lp_arena: &mut Arena<IR>) -> LogicalPlan {
-    let alp = lp_arena.get_mut(node);
-    let alp = std::mem::take(alp);
-    alp.into_lp(
-        &|node, lp_arena: &mut Arena<IR>| {
-            let lp = lp_arena.get_mut(node);
-            std::mem::take(lp)
-        },
-        lp_arena,
-        expr_arena,
-    )
 }
